Fix #15 (Piracy terms check seems to malfunction) (#17)
Attempt to fix #15. The strings are escaped upon initialization before being added to the regex. Previously all periods were interpreted as "any character". That's obviously not what we want.

The escapeRegExp function could or should probably be moved somewhere else, but I guess for now it should be fine where it is.